fix: Fix bug related to placing content blocks
Content Blocks cannot be placed correctly when new rows are added.
This is fixed by improving the handling of the rowCount state for
each layout. A part of the solution is to remove the minRows state.
This state can be reimplemented in a way that is compatiable with the
current changes, if needed.